Added rope changes
